Remove incorrect option passed to esbuild minifier
Newer esbuild versions validate the options passed to it's JS api, and this option was accidentally being passed. `devtool` is a webpack thing, not an `esbuild` thing, and we just use it to decide if we should ask `esbuild` for sourcemaps. Removing this option fixes warnings when using the `esbuild-loader` with `esbuild` 0.7.1.